Fix some compile warnings & errors
* Fix type of literals
* Add explicit casts
* #include missing headers
* Replace use of non-standard u_int8_t
* Remove name of unused variables
* Fix DLL export macro - this was broken when some of the exporting headers
were moved to a different library.

+++ src/backends/reference/workloads/BaseIterator.hpp
@@ -456,7 +456,7 @@ public:
         for (uint32_t i = 0; i < size; ++i)
         {
             this->operator[](i);
-            decodedTensor.emplace_back(*m_Iterator);
+            decodedTensor.emplace_back(static_cast<float>(*m_Iterator));
         }
 
         return decodedTensor;
